+++ src/Servers/Linux.php
@@ -17,6 +17,7 @@ use Innmind\Server\Status\{
 use Innmind\Server\Control\Server as Control;
 use Innmind\TimeContinuum\Clock;
 use Innmind\Url\Path;
+use Innmind\Immutable\Attempt;
 use Innmind\Immutable\Maybe;
 
 final class Linux implements Server
@@ -37,7 +38,7 @@ final class Linux implements Server
     }
 
     #[\Override]
-    public function cpu(): Maybe
+    public function cpu(): Attempt
     {
         return ($this->cpu)();
     }